I ADDED AN EMPTY HOMEPAGE COMPONENT IN PAGES
I added the pages folder in src and added a new component named
HomePage. This will be the home page.

I added a test for the HomePage, HomePage-component.spec.jsx, for
testing the home page. It simply tests at this point that the
component does not throw an exception or error.

App.spec.jsx now has an empty test because App.jsx now has a route to
the home page.

I added jsconfig.json to add intellisense for Jest. It worked for the
App test file but not the new HomePage component test file. I found
this blog post which solved my problem:
https://medium.com/@chris.washington_60485/virtual-studio-code-and-jest-intellisense-headaches-cured-5425ed914f43