I HAVE A WORKING HEADER BAR
In package.json, I reinstalled enzyme and the react 16 enzyme adapter
as dev dependencies, not production dependencies.

App.jsx has been modified to use react router. I thought that I
included it as part of the previous commit but apparently I did not.
It is now has a single route to home page.

In index.scss, I set using the box-sizing to border-box using the
universal selector , and set the font-size to 16px so that I could
use the REM system for specifying sizes.

App.spec.jsx had a typo in its header; I called the file “.js”. This
has been changed to “.jsx”.

I added the components folder and added the header component. I want
to get that working first. This is also the first component that I have
styled. I also included the images for GitHub, Linked In, and Medium
 where I shall blog. So, there is a portfolio link, and links for
 Github, LinkedIn and Medium.