I HAVE A BASIC HOMEPAGE WITH MY TYPING TEXT
I had to remove the exact path for the homepage. It caused the HomePage
component to not route and render.

I removed the <Link> in the header component since it was causing an
error in the console.

My homepage now has the scrolling text like https://caferati.me. This
will need to become a component in my next commit.