TYPING TEXT IS WORKING AS A COMPONENT
I took the code that was in HomePage and moved it to a new component,
TypingText-component.jsx. It takes a string of text to render and it
takes an interval timer in seconds.

HomePage is now using the component. I added some styling to HomePage.

I changed the background color in index.scss to match caferati.me.