I AM FIXING MEMORY LEAKS
I saw in the Chrome console that I am receiving the following error:

Warning: Can't perform a React state update on an unmounted component.
This is a no-op, but it indicates a memory leak in your application.
To fix, cancel all subscriptions and asynchronous tasks in the
componentWillUnmount method.
    in HomePage (created by Context.Consumer)

I now clear the timers but that did not fix anything. When I comment
the TypingText component and switch from HomePage to Skills, the
message no longer appears. But the TypingText component should know how
to cleanup itself. I have worked on this 5 hours and must move on to
somethig else.