MASSIVE REFACOTRING BASED ON PAUL'S FEEDBACK
Paul gave me feedback last night.

There is more padding in the header. I also lowered the "yellow bar"
that marks the bottom of the header.

The section dividers are now thinner.

The big thing was learning about scroll-margin-top. This makes it
easier to have the # links not scroll under the header. This is built
for that and I am using it and it made life easier.

I found an arrow image and added it to the bottom of the page. It takes
you back to the top of the page.

There are other minor tweaks.